refactor(kanban): break into components, fix drag cancel and won status
