+++ src/app/(main)/dashboard/kanban/_components/data.ts
@@ -1,0 +1,247 @@
+import type { BoardState, DealTag } from "./types";
+
+export const tagTones: Record<DealTag, string> = {
+  Analytics: "bg-cyan-500/10 text-cyan-700 dark:text-cyan-300",
+  Automation: "bg-violet-500/10 text-violet-700 dark:text-violet-300",
+  Branding: "bg-orange-500/10 text-orange-700 dark:text-orange-300",
+  Consulting: "bg-red-500/10 text-red-700 dark:text-red-300",
+  CRM: "bg-sky-500/10 text-sky-700 dark:text-sky-300",
+  Design: "bg-violet-500/10 text-violet-700 dark:text-violet-300",
+  Dev: "bg-amber-500/10 text-amber-700 dark:text-amber-300",
+  "E-commerce": "bg-emerald-500/10 text-emerald-700 dark:text-emerald-300",
+  Enterprise: "bg-red-500/10 text-red-700 dark:text-red-300",
+  Growth: "bg-orange-500/10 text-orange-700 dark:text-orange-300",
+  Integration: "bg-violet-500/10 text-violet-700 dark:text-violet-300",
+  IT: "bg-blue-500/10 text-blue-700 dark:text-blue-300",
+  Marketing: "bg-teal-500/10 text-teal-700 dark:text-teal-300",
+  Mobile: "bg-blue-500/10 text-blue-700 dark:text-blue-300",
+  Platform: "bg-violet-500/10 text-violet-700 dark:text-violet-300",
+  Software: "bg-blue-500/10 text-blue-700 dark:text-blue-300",
+  Web: "bg-blue-500/10 text-blue-700 dark:text-blue-300",
+};
+
+export const initialBoard: BoardState = {
+  lead: [
+    {
+      id: "orion-labs",
+      company: "Orion Labs",
+      description: "Website redesign",
+      amount: 24_000,
+      date: "May 16",
+      tags: ["Web", "Design"],
+      comments: 2,
+      files: 1,
+      logo: "O",
+      logoTone: "bg-indigo-950 text-white dark:bg-indigo-400 dark:text-indigo-950",
+      assignee: "AM",
+      assigneeTone: "bg-amber-100 text-amber-950 dark:bg-amber-300",
+    },
+    {
+      id: "pioneer-foods",
+      company: "Pioneer Foods",
+      description: "E-commerce platform",
+      amount: 18_500,
+      date: "May 20",
+      tags: ["E-commerce", "Dev"],
+      comments: 1,
+      files: 1,
+      logo: "P",
+      logoTone: "bg-green-500 text-white",
+      assignee: "DR",
+      assigneeTone: "bg-orange-100 text-orange-950 dark:bg-orange-300",
+    },
+    {
+      id: "summit-realty",
+      company: "Summit Realty",
+      description: "CRM implementation",
+      amount: 15_000,
+      date: "May 28",
+      tags: ["CRM", "Consulting"],
+      comments: 0,
+      files: 2,
+      logo: "S",
+      logoTone: "bg-violet-500/10 text-violet-700 dark:text-violet-300",
+      assignee: "JL",
+      assigneeTone: "bg-yellow-100 text-yellow-950 dark:bg-yellow-300",
+    },
+  ],
+  qualified: [
+    {
+      id: "bluewave-solutions",
+      company: "Bluewave Solutions",
+      description: "Marketing automation setup",
+      amount: 32_000,
+      date: "May 22",
+      tags: ["Marketing", "Automation"],
+      comments: 2,
+      files: 3,
+      logo: "B",
+      logoTone: "bg-sky-500/10 text-sky-700 dark:text-sky-300",
+      assignee: "NK",
+      assigneeTone: "bg-stone-200 text-stone-950 dark:bg-stone-300",
+    },
+    {
+      id: "nova-tech",
+      company: "Nova Tech",
+      description: "Mobile app development",
+      amount: 45_000,
+      date: "May 30",
+      tags: ["Mobile", "Dev"],
+      comments: 1,
+      files: 2,
+      logo: "N",
+      logoTone: "bg-neutral-950 text-white dark:bg-neutral-100 dark:text-neutral-950",
+      assignee: "AM",
+      assigneeTone: "bg-amber-100 text-amber-950 dark:bg-amber-300",
+    },
+    {
+      id: "brightly",
+      company: "Brightly",
+      description: "Brand identity design",
+      amount: 9_500,
+      date: "Jun 2",
+      tags: ["Design", "Branding"],
+      comments: 0,
+      files: 1,
+      logo: "B",
+      logoTone: "bg-amber-500/10 text-amber-700 dark:text-amber-300",
+      assignee: "DR",
+      assigneeTone: "bg-orange-100 text-orange-950 dark:bg-orange-300",
+    },
+    {
+      id: "greenfield-inc",
+      company: "Greenfield Inc.",
+      description: "IT infrastructure upgrade",
+      amount: 70_000,
+      date: "Jun 5",
+      tags: ["IT", "Consulting"],
+      comments: 1,
+      files: 4,
+      logo: "G",
+      logoTone: "bg-green-500/10 text-green-700 dark:text-green-300",
+      assignee: "NK",
+      assigneeTone: "bg-stone-200 text-stone-950 dark:bg-stone-300",
+    },
+  ],
+  proposal: [
+    {
+      id: "techsphere",
+      company: "TechSphere",
+      description: "Custom software solution",
+      amount: 75_000,
+      date: "Jun 10",
+      tags: ["Software", "Dev"],
+      comments: 2,
+      files: 3,
+      logo: "T",
+      logoTone: "bg-blue-600 text-white",
+      assignee: "AM",
+      assigneeTone: "bg-amber-100 text-amber-950 dark:bg-amber-300",
+    },
+    {
+      id: "lumen-corp",
+      company: "Lumen Corp",
+      description: "Data analytics platform",
+      amount: 50_000,
+      date: "Jun 12",
+      tags: ["Analytics", "Platform"],
+      comments: 1,
+      files: 2,
+      logo: "L",
+      logoTone: "bg-indigo-600 text-white",
+      assignee: "NK",
+      assigneeTone: "bg-stone-200 text-stone-950 dark:bg-stone-300",
+    },
+    {
+      id: "wellness-co",
+      company: "Wellness Co.",
+      description: "Customer portal development",
+      amount: 50_000,
+      date: "Jun 15",
+      tags: ["Web", "Dev"],
+      comments: 0,
+      files: 1,
+      logo: "W",
+      logoTone: "bg-emerald-500/10 text-emerald-700 dark:text-emerald-300",
+      assignee: "JL",
+      assigneeTone: "bg-yellow-100 text-yellow-950 dark:bg-yellow-300",
+    },
+  ],
+  negotiation: [
+    {
+      id: "apex-industries",
+      company: "Apex Industries",
+      description: "Manufacturing CRM",
+      amount: 60_000,
+      date: "Jun 20",
+      tags: ["CRM", "Enterprise"],
+      comments: 2,
+      files: 4,
+      logo: "A",
+      logoTone: "bg-blue-500/10 text-blue-700 dark:text-blue-300",
+      assignee: "NK",
+      assigneeTone: "bg-stone-200 text-stone-950 dark:bg-stone-300",
+    },
+    {
+      id: "velocity-partners",
+      company: "Velocity Partners",
+      description: "Lead generation campaign",
+      amount: 50_000,
+      date: "Jun 18",
+      tags: ["Marketing", "Growth"],
+      comments: 1,
+      files: 2,
+      logo: "V",
+      logoTone: "bg-orange-500/10 text-orange-700 dark:text-orange-300",
+      assignee: "JL",
+      assigneeTone: "bg-yellow-100 text-yellow-950 dark:bg-yellow-300",
+    },
+  ],
+  won: [
+    {
+      id: "horizon-ltd",
+      company: "Horizon Ltd.",
+      description: "Website & branding package",
+      amount: 45_000,
+      date: "May 2",
+      tags: ["Web", "Branding"],
+      comments: 0,
+      files: 0,
+      logo: "H",
+      logoTone: "bg-sky-600 text-white",
+      assignee: "NK",
+      assigneeTone: "bg-stone-200 text-stone-950 dark:bg-stone-300",
+      won: true,
+    },
+    {
+      id: "cloudsync",
+      company: "CloudSync",
+      description: "SaaS integration project",
+      amount: 80_000,
+      date: "Apr 28",
+      tags: ["Integration", "Dev"],
+      comments: 0,
+      files: 0,
+      logo: "C",
+      logoTone: "bg-blue-500/10 text-blue-700 dark:text-blue-300",
+      assignee: "AM",
+      assigneeTone: "bg-amber-100 text-amber-950 dark:bg-amber-300",
+      won: true,
+    },
+    {
+      id: "finedge",
+      company: "FinEdge",
+      description: "Financial dashboard",
+      amount: 80_000,
+      date: "Apr 15",
+      tags: ["Analytics", "Platform"],
+      comments: 0,
+      files: 0,
+      logo: "F",
+      logoTone: "bg-emerald-500/10 text-emerald-700 dark:text-emerald-300",
+      assignee: "DR",
+      assigneeTone: "bg-orange-100 text-orange-950 dark:bg-orange-300",
+      won: true,
+    },
+  ],
+};
